requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
alembic
psycopg2
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==4.3.0
bleach
httpx
pytest
